# perf(metarepos): add a pool for []*mrpb.Report