@ijsong 2023-07-18
Current dependencies on/for this PR:
* main
  * **PR #532** <a href="https://app.graphite.dev/github/pr/kakao/varlog/532" target="_blank"><img src="https://static.graphite.dev/graphite-32x32.png" alt="Graphite" width="10px" height="10px"/></a> 
    * **PR #533** <a href="https://app.graphite.dev/github/pr/kakao/varlog/533" target="_blank"><img src="https://static.graphite.dev/graphite-32x32.png" alt="Graphite" width="10px" height="10px"/></a> 
      * **PR #534** <a href="https://app.graphite.dev/github/pr/kakao/varlog/534" target="_blank"><img src="https://static.graphite.dev/graphite-32x32.png" alt="Graphite" width="10px" height="10px"/></a>  👈

This comment was auto-generated by [Graphite](https://app.graphite.dev/github/pr/kakao/varlog/534?utm_source=stack-comment).