@ijsong 2023-07-28
@ijsong started a stack merge that includes this pull request via [Graphite](https://app.graphite.dev/github/pr/kakao/varlog/534).